+++ src/main/resources/ValidationMessages.properties
@@ -1,5 +1,5 @@
+javax.validation.constraints.AssertTrue.message=チェックしてください。
 javax.validation.constraints.Email.message=正しいメールアドレスを入力してください。
 javax.validation.constraints.NotBlank.message=入力してください。
 javax.validation.constraints.NotNull.message=入力してください。
 javax.validation.constraints.Size.message={min} 文字以上 {max} 文字以下で入力してください。
-vec.presentation.form.account.DeleteForm.confirmation.AssertTrue.message=チェックしてください。
